styling
I styled my website to be displayed on mobile screen devices. In my HTML, I removed extra breaks between form and image and placed my button inside a separate div but it's still a part of the form. In my stylesheet, I changed the positions of my elements, added and changed colours and added a hover effect to my button.